move RegistryConfig to resolver package
This allows using the resolver package without having to import
the buildkit daemon configuration.
